+++ fiel-back/src/main/java/com/mahas/dto/request/sale/SaleDTORequest.java
@@ -1,5 +1,6 @@
 package com.mahas.dto.request.sale;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.mahas.dto.request.DTORequest;
@@ -16,7 +17,7 @@ import lombok.Setter;
 public class SaleDTORequest implements DTORequest {
     private Integer id;
     private Integer user;
-    private Double freight;
+    private BigDecimal freight;
     private LocalDate deliveryDate;
     private Integer status;
     private Integer address;
